[SettingsMigration] Move setting migration step primitives into own .cpp/.h
