+++ xbmc/settings/lib/CMakeLists.txt
@@ -8,6 +8,7 @@ set(SOURCES ISetting.cpp
             SettingSection.cpp
             SettingsManager.cpp
             SettingsMigration.cpp
+            SettingsMigrationPrimitive.cpp
             SettingsMigrationSteps.cpp
             SettingUpdate.cpp)
 
@@ -29,6 +30,7 @@ set(HEADERS ISetting.h
             SettingSection.h
             SettingsManager.h
             SettingsMigration.h
+            SettingsMigrationPrimitive.h
             SettingsMigrationSteps.h
             SettingType.h
             SettingUpdate.h)
