+++ xbmc/settings/lib/test/CMakeLists.txt
@@ -1,4 +1,5 @@
 set(SOURCES TestSettingsManager.cpp
-            TestSettingsMigration.cpp)
+            TestSettingsMigration.cpp
+            TestSettingsMigrationPrimitive.cpp)
 
 core_add_test_library(settings_lib_test)
